+++ rewrite/changelog.md
@@ -20,6 +20,7 @@ The subsections are formatted as they need when publishing a release to the Play
 ## Unreleased
 
  * Reduced log file max size to 256K
+ * Explicit wakelock
 
 ## [0.9.13] - 2019-08-17 13: